backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
aiosqlite==0.19.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
playwright==1.40.0
openai==1.3.7
python-docx==1.1.0
PyPDF2==3.0.1
beautifulsoup4==4.12.2
lxml==4.9.3
aiofiles==23.2.1 